fix: relax channel requirement
We relax the channel requirement from global to local. That is, channels in the same rank must have different send peers (recv peers) and we allow the number of threadblocks to exceed 78 in a rank.
Correspondingly, we adjust the CPU thread launch logic to pick 78 threadblocks randomly at each iteration. Other threads can only start when one of them entires. This simulates the GPU execution model of Nvidia H20 GPUs.